Mahendra_Kumar_Reddy_Kakarla/contractreviewassistant/requirements.txt
PyPDF2
python-docx
google-generativeai
pyyaml
asyncio
langchain
langchain_core
langchain_google_genai
python-dotenv